+++ pkg/exporter/survey-utils.go
@@ -30,7 +30,9 @@ func extractQuestions(root *types.SurveyItem, prefLang string, includeItemNames 
 		if item.Type == "pageBreak" {
 			continue
 		}
-
+		if item.ConfidentialMode != "" {
+			continue
+		}
 		if isItemGroup(&item) {
 			questions = append(questions, extractQuestions(&item, prefLang, includeItemNames, excludeItemNames)...)
 			continue
@@ -205,6 +207,10 @@ func mapToResponseDef(rItem *types.ItemComponent, parentKey string, lang string)
 				option.OptionType = OPTION_TYPE_CLOZE
 			}
 			responseDef.Options = append(responseDef.Options, option)
+			if option.OptionType == OPTION_TYPE_CLOZE {
+				clozeOptions := extractClozeInputOptions(o, option.ID, lang)
+				responseDef.Options = append(responseDef.Options, clozeOptions...)
+			}
 		}
 		responseDef.ResponseType = QUESTION_TYPE_SINGLE_CHOICE
 		return []ResponseDef{responseDef}
@@ -233,6 +239,10 @@ func mapToResponseDef(rItem *types.ItemComponent, parentKey string, lang string)
 				option.OptionType = OPTION_TYPE_CLOZE
 			}
 			responseDef.Options = append(responseDef.Options, option)
+			if option.OptionType == OPTION_TYPE_CLOZE {
+				clozeOptions := extractClozeInputOptions(o, option.ID, lang)
+				responseDef.Options = append(responseDef.Options, clozeOptions...)
+			}
 		}
 		responseDef.ResponseType = QUESTION_TYPE_MULTIPLE_CHOICE
 		return []ResponseDef{responseDef}
@@ -604,7 +614,9 @@ func mapToResponseDef(rItem *types.ItemComponent, parentKey string, lang string)
 			case "dropDownGroup":
 				option.OptionType = OPTION_TYPE_DROPDOWN
 			}
-			responseDef.Options = append(responseDef.Options, option)
+			if option.OptionType != "" {
+				responseDef.Options = append(responseDef.Options, option)
+			}
 		}
 		responseDef.ResponseType = QUESTION_TYPE_CLOZE
 		return []ResponseDef{responseDef}
@@ -642,4 +654,34 @@ func getQuestionType(responses []ResponseDef) string {
 	}
 
 	return qType
+}
+
+func extractClozeInputOptions(option types.ItemComponent, clozeKey string, lang string) (clozeoptions []ResponseOption) {
+	clozeInputs := []ResponseOption{}
+	for _, o := range option.Items {
+		label, err := getPreviewText(&o, lang)
+		if err != nil {
+			logger.Debug.Printf("mapToResponseDef: label not found for: %v", o)
+		}
+		option := ResponseOption{
+			ID:    clozeKey + "." + o.Key,
+			Label: label,
+		}
+		switch o.Role {
+		case "input":
+			option.OptionType = OPTION_TYPE_EMBEDDED_CLOZE_TEXT_INPUT
+		case "dateInput":
+			option.OptionType = OPTION_TYPE_EMBEDDED_CLOZE_DATE_INPUT
+		case "timeInput":
+			option.OptionType = OPTION_TYPE_EMBEDDED_CLOZE_NUMBER_INPUT
+		case "numberInput":
+			option.OptionType = OPTION_TYPE_EMBEDDED_CLOZE_NUMBER_INPUT
+		case "dropDownGroup":
+			option.OptionType = OPTION_TYPE_EMBEDDED_CLOZE_DROPDOWN
+		}
+		if option.OptionType != "" {
+			clozeInputs = append(clozeInputs, option)
+		}
+	}
+	return clozeInputs
 }