+++ pkg/types/participant-state.go
@@ -6,9 +6,10 @@ import (
 )
 
 const (
-	PARTICIPANT_STUDY_STATUS_ACTIVE    = "active"
-	PARTICIPANT_STUDY_STATUS_TEMPORARY = "temporary" // for participants without a registered account
-	PARTICIPANT_STUDY_STATUS_EXITED    = "exited"
+	PARTICIPANT_STUDY_STATUS_ACTIVE          = "active"
+	PARTICIPANT_STUDY_STATUS_TEMPORARY       = "temporary" // for participants without a registered account
+	PARTICIPANT_STUDY_STATUS_EXITED          = "exited"
+	PARTICIPANT_STUDY_STATUS_ACCOUNT_DELETED = "accountDeleted"
 )
 
 // ParticipantState defines the datamodel for current state of the participant in a study as stored in the database